requirements.txt
django
djangorestframework
celery
redis
gunicorn